+++ src/test/java/seedu/address/logic/parser/ArgumentMultimapTest.java
@@ -4,12 +4,15 @@ import static org.junit.Assert.assertEquals;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_AMY;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
 import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_NAME;
+import static seedu.address.logic.parser.util.CliSyntax.PREFIX_PHONE;
 
 import java.util.HashMap;
 
 import org.junit.Test;
+
+import seedu.address.logic.parser.util.ArgumentMultimap;
+import seedu.address.logic.parser.util.Prefix;
 
 //@@author yunpengn
 public class ArgumentMultimapTest {
